refactor(api): restrukturisasi ke pola controller dan router
- Memindahkan logika bisnis dari index.ts ke folder controllers

- Implementasi modular routing di folder routes

- Pembersihan entry point index.ts untuk skalabilitas
