+++ apps/api/src/controllers/orderController.ts
@@ -1,0 +1,330 @@
+import { FastifyRequest, FastifyReply } from "fastify";
+import { z } from "zod";
+import { prisma } from "../prisma";
+
+// ====== 1. SCHEMA VALIDASI (ZOD) ======
+const orderItemSchema = z.object({
+  menuId: z.number().int().positive(),
+  qty: z.number().int().positive(),
+});
+
+const orderCreateSchema = z.object({
+  origin: z.enum(["COUNTER", "KITCHEN", "BAR"]),
+  customerName: z.string().trim().min(1).optional(),
+  items: z.array(orderItemSchema).min(1),
+});
+
+const paySchema = z.object({
+  paymentMethod: z.enum(["CASH", "QRIS"]),
+  overrideStock: z.boolean().optional().default(false),
+  overrideNote: z.string().trim().optional(),
+});
+
+// ====== 2. HELPER FUNGSI STOK ======
+type StockRequirement = {
+  ingredientId: number;
+  ingredientName: string;
+  unit: string;
+  stock: number;
+  need: number;
+  shortBy: number;
+};
+
+function buildStockRequirements(
+  details: Array<{ menuId: number; qty: number }>,
+  recipes: Array<{ menuId: number; ingredientId: number; amountNeeded: number }>,
+  ingredients: Array<{ id: number; name: string; stock: number; unit: string }>
+): StockRequirement[] {
+  const qtyByMenu = new Map<number, number>();
+
+  for (const d of details) {
+    qtyByMenu.set(d.menuId, (qtyByMenu.get(d.menuId) ?? 0) + d.qty);
+  }
+
+  const totalNeeded = new Map<number, number>();
+
+  for (const r of recipes) {
+    const qty = qtyByMenu.get(r.menuId) ?? 0;
+    const need = r.amountNeeded * qty;
+    totalNeeded.set(r.ingredientId, (totalNeeded.get(r.ingredientId) ?? 0) + need);
+  }
+
+  const ingredientMap = new Map(ingredients.map((item) => [item.id, item]));
+
+  return [...totalNeeded.entries()].map(([ingredientId, need]) => {
+    const ingredient = ingredientMap.get(ingredientId);
+
+    if (!ingredient) {
+      throw new Error(`Ingredient ${ingredientId} not found`);
+    }
+
+    const shortBy = Math.max(need - ingredient.stock, 0);
+
+    return {
+      ingredientId,
+      ingredientName: ingredient.name,
+      unit: ingredient.unit,
+      stock: ingredient.stock,
+      need,
+      shortBy,
+    };
+  });
+}
+
+async function getOrderStockSummary(orderId: number) {
+  const order = await prisma.order.findUnique({
+    where: { id: orderId },
+    include: { details: true },
+  });
+
+  if (!order) return null;
+
+  const menuIds = [...new Set(order.details.map((d) => d.menuId))];
+
+  const recipes = menuIds.length
+    ? await prisma.recipe.findMany({
+        where: { menuId: { in: menuIds } },
+      })
+    : [];
+
+  const ingredientIds = [...new Set(recipes.map((r) => r.ingredientId))];
+
+  const ingredients = ingredientIds.length
+    ? await prisma.ingredient.findMany({
+        where: { id: { in: ingredientIds } },
+      })
+    : [];
+
+  const requirements = buildStockRequirements(order.details, recipes, ingredients);
+  const shortages = requirements.filter((item) => item.shortBy > 0);
+
+  return {
+    order,
+    requirements,
+    shortages,
+  };
+}
+
+// ====== 3. FUNGSI CONTROLLER ======
+
+export const createOrder = async (req: FastifyRequest, reply: FastifyReply) => {
+  const parsed = orderCreateSchema.safeParse(req.body);
+
+  if (!parsed.success) {
+    return reply.code(400).send({ error: parsed.error.flatten() });
+  }
+
+  const userId = (req as any).user?.id || (req as any).user?.userId;
+
+  if (!userId) {
+    return reply.code(401).send({ error: "Unauthorized: User ID tidak ditemukan dalam token." });
+  }
+
+  const { origin, customerName, items } = parsed.data;
+
+  const menus = await prisma.menu.findMany({
+    where: { id: { in: items.map((i) => i.menuId) } },
+  });
+
+  const menuMap = new Map(menus.map((m) => [m.id, m]));
+
+  if (menus.length !== new Set(items.map((i) => i.menuId)).size) {
+    return reply.code(400).send({ error: "Ada menu yang tidak ditemukan." });
+  }
+
+  const details = items.map((i) => {
+    const menu = menuMap.get(i.menuId)!;
+    const subtotal = menu.price * i.qty;
+
+    return {
+      menuId: menu.id,
+      qty: i.qty,
+      price: menu.price,
+      subtotal,
+    };
+  });
+
+  const totalPrice = details.reduce((sum, item) => sum + item.subtotal, 0);
+
+  const order = await prisma.order.create({
+    data: {
+      userId: Number(userId),
+      origin,
+      customerName,
+      totalPrice,
+      paymentStatus: "UNPAID",
+      details: {
+        create: details,
+      },
+    },
+    include: { details: true },
+  });
+
+  return reply.send(order);
+};
+
+export const checkOrderStock = async (req: FastifyRequest, reply: FastifyReply) => {
+  const id = Number((req.params as any).id);
+
+  if (!Number.isFinite(id)) {
+    return reply.code(400).send({ error: "Invalid id" });
+  }
+
+  const summary = await getOrderStockSummary(id);
+
+  if (!summary) {
+    return reply.code(404).send({ error: "Order not found" });
+  }
+
+  return reply.send({
+    ok: true,
+    orderId: summary.order.id,
+    hasShortage: summary.shortages.length > 0,
+    requirements: summary.requirements,
+    shortages: summary.shortages,
+  });
+};
+
+export const payOrder = async (req: FastifyRequest, reply: FastifyReply) => {
+  const parsed = paySchema.safeParse(req.body);
+
+  if (!parsed.success) {
+    return reply.code(400).send({ error: parsed.error.flatten() });
+  }
+
+  const { paymentMethod, overrideStock, overrideNote } = parsed.data;
+  const id = Number((req.params as any).id);
+
+  if (!Number.isFinite(id)) {
+    return reply.code(400).send({ error: "Invalid id" });
+  }
+
+  const txResult = await prisma.$transaction(async (tx) => {
+    await tx.$executeRawUnsafe(
+      `SELECT id FROM "Order" WHERE id = ${id} FOR UPDATE`
+    );
+
+    const lockedOrder = await tx.order.findUnique({
+      where: { id },
+      include: { details: true },
+    });
+
+    if (!lockedOrder) {
+      return { kind: "NOT_FOUND" as const };
+    }
+
+    if (lockedOrder.paymentStatus === "PAID") {
+      return {
+        kind: "ALREADY_PAID" as const,
+        order: lockedOrder,
+      };
+    }
+
+    if (lockedOrder.paymentStatus === "VOID") {
+      return { kind: "VOID" as const };
+    }
+
+    const menuIds = [...new Set(lockedOrder.details.map((d) => d.menuId))];
+
+    const recipes = menuIds.length
+      ? await tx.recipe.findMany({
+          where: { menuId: { in: menuIds } },
+        })
+      : [];
+
+    const ingredientIds = [...new Set(recipes.map((r) => r.ingredientId))];
+
+    if (ingredientIds.length) {
+      await tx.$executeRawUnsafe(
+        `SELECT id FROM "Ingredient" WHERE id IN (${ingredientIds.join(",")}) FOR UPDATE`
+      );
+    }
+
+    const lockedIngredients = ingredientIds.length
+      ? await tx.ingredient.findMany({
+          where: { id: { in: ingredientIds } },
+        })
+      : [];
+
+    const requirements = buildStockRequirements(
+      lockedOrder.details,
+      recipes,
+      lockedIngredients
+    );
+
+    const shortages = requirements.filter((item) => item.shortBy > 0);
+
+    if (shortages.length > 0 && !overrideStock) {
+      return {
+        kind: "SHORTAGE" as const,
+        shortages,
+      };
+    }
+
+    const didOverride = shortages.length > 0 && overrideStock;
+
+    const paidOrder = await tx.order.update({
+      where: { id: lockedOrder.id },
+      data: {
+        paymentStatus: "PAID",
+        paymentMethod,
+        stockOverride: didOverride,
+        overrideNote: didOverride ? overrideNote : null,
+        paidAt: lockedOrder.paidAt ?? new Date(),
+      },
+      include: { details: true },
+    });
+
+    for (const item of requirements) {
+      await tx.ingredient.update({
+        where: { id: item.ingredientId },
+        data: { stock: { decrement: item.need } },
+      });
+
+      await tx.stockMovement.create({
+        data: {
+          ingredientId: item.ingredientId,
+          qtyChange: -item.need,
+          reason: didOverride ? "SALE_OVERRIDE" : "SALE",
+          orderId: lockedOrder.id,
+        },
+      });
+    }
+
+    return {
+      kind: "PAID" as const,
+      order: paidOrder,
+      shortages,
+    };
+  });
+
+  if (txResult.kind === "NOT_FOUND") {
+    return reply.code(404).send({ error: "Order not found" });
+  }
+
+  if (txResult.kind === "VOID") {
+    return reply.code(409).send({ error: "Order already VOID" });
+  }
+
+  if (txResult.kind === "SHORTAGE") {
+    return reply.code(409).send({
+      error: "STOCK_NOT_ENOUGH",
+      shortages: txResult.shortages,
+    });
+  }
+
+  if (txResult.kind === "ALREADY_PAID") {
+    return reply.send({
+      ok: true,
+      alreadyPaid: true,
+      order: txResult.order,
+    });
+  }
+
+  return reply.send({
+    ok: true,
+    alreadyPaid: false,
+    order: txResult.order,
+    shortages: txResult.shortages,
+  });
+};
